requirements.txt
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
cycler==0.10.0
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pycparser==2.20
pyparsing==2.4.7
pyperclip==1.8.1
python-dateutil==2.8.1
python-decouple==3.4
six==1.15.0
texttable==1.6.3
